requirement.txt
pytorch_wavelets
thop
PyWavelets
opencv-python
scikit-learn
seaborn
tqdm
matplotlib